refactor(vx-scan): Control sounds using query change listener
Rather than playing sounds when screens mount using useEffect, add in a logic handler that plays sounds according to scanner status transitions. This decouples sounds from how/when we mount and render screens, making them easier to grok and less error-prone.
