--- apps/vx-scan/frontend/src/hooks/use_sound.ts
@@ -1,13 +1,0 @@
-import { useContext } from 'react';
-import useSoundLib from 'use-sound';
-import { AppContext } from '../contexts/app_context';
-
-export function useSound(sound: 'success' | 'warning' | 'error'): () => void {
-  const [playSound] = useSoundLib(`/sounds/${sound}.mp3`);
-  const { isSoundMuted } = useContext(AppContext);
-  if (isSoundMuted) {
-    // eslint-disable-next-line @typescript-eslint/no-empty-function
-    return () => {};
-  }
-  return playSound;
-}
