+++ apps/vx-scan/frontend/src/screens/scan_busy_screen.tsx
@@ -1,16 +1,12 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { Text } from '@votingworks/ui';
 import { ExclamationTriangle } from '../components/graphics';
 import {
   CenteredLargeProse,
   ScreenMainCenterChild,
 } from '../components/layout';
-import { useSound } from '../hooks/use_sound';
 
 export function ScanBusyScreen(): JSX.Element {
-  const playWarning = useSound('warning');
-  useEffect(playWarning, [playWarning]);
-
   return (
     <ScreenMainCenterChild infoBar={false}>
       <ExclamationTriangle />
