+++ apps/vx-scan/frontend/src/screens/scan_success_screen.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { CircleCheck } from '../components/graphics';
 import { ScannedBallotCount } from '../components/scanned_ballot_count';
 
@@ -6,16 +6,12 @@ import {
   CenteredLargeProse,
   ScreenMainCenterChild,
 } from '../components/layout';
-import { useSound } from '../hooks/use_sound';
 
 interface Props {
   scannedBallotCount: number;
 }
 
 export function ScanSuccessScreen({ scannedBallotCount }: Props): JSX.Element {
-  const playSuccess = useSound('success');
-  useEffect(playSuccess, [playSuccess]);
-
   return (
     <ScreenMainCenterChild>
       <CircleCheck />
