+++ apps/vx-scan/frontend/src/utils/use_sound.ts
@@ -1,0 +1,6 @@
+import useSoundLib from 'use-sound';
+
+export function useSound(sound: 'success' | 'warning' | 'error'): () => void {
+  const [playSound] = useSoundLib(`/sounds/${sound}.mp3`);
+  return playSound;
+}
